The Door Lock Project
This is an Arduino Project that utilizes the 4x4 Keypad and the stepper motor. The arduino receives an input from the user from a decided password labeled as unlockPassword and lockPassword and will decide whether to lock, unlock or to do neither. The program will indicate when ready to beign and will either return "Correct, Unlocking", "Correct, Locking" or "incorrect Password"