Vending Machine Application
This is a Java Application that has a list of items that can be interchanged or added with a code that mimics modern vending machines. The vending machine also has a bank system that checks the value of each coins to return to the user and will keep a tab or how much was gained, returned and total that was made during a program run. The program also has a developer tool that can help owner see the amounts of sodas and coins that was used for debugging.